# [IMP] core: speed up registry setup