--- a/odoo/addons/base/models/ir_rule.py
+++ b/odoo/addons/base/models/ir_rule.py
@@ -280,6 +280,7 @@ class IrRule(models.Model):
 # 'global' is a Python keyword. Therefore, we add it to the class by assignment.
 # Note that the attribute '_module' is normally added by the class' metaclass.
 #
-setattr(IrRule, 'global',
-        fields.Boolean(compute='_compute_global', store=True, _module=IrRule._module,
-                       help="If no group is specified the rule is global and applied to everyone"))
+global_ = fields.Boolean(compute='_compute_global', store=True,
+                         help="If no group is specified the rule is global and applied to everyone")
+setattr(IrRule, 'global', global_)
+global_.__set_name__(IrRule, 'global')