… addons/website/tests/test_performance.py
… addons/website_livechat/tests/test_livechat_basic_flow.py
… odoo/addons/base/tests/test_ir_actions.py
… odoo/addons/base/tests/test_ir_model.py
… odoo/addons/test_inherit/tests/test_inherit.py
… odoo/addons/test_inherits/tests/test_inherits.py
… odoo/addons/test_new_api/tests/test_new_fields.py
… odoo/fields.py
… odoo/models.py
… odoo/tests/common.py
… odoo/tools/misc.py