@fmdl 2021-04-27
@rco-odoo can you explain what is the new attribute `recursive` ?