@Elkasitu 2021-04-28
> @rco-odoo can you explain what is the new attribute `recursive` ?
A recursive field is a compute field that depends on itself in some capacity in order to perform the computation.

It's not really new in the sense that the attribute existed prior to this patch, the difference was that it was dynamically set (the ORM did the job of figuring out which fields were recursive and which ones weren't) and now it's done statically (when defining the field). The reason why it's done statically now is because doing it dynamically would rely on components that make the sharing of the field impossible (which reduces bootstrapping performance)