@Xavier-Do 2021-04-28
> > @rco-odoo can you explain what is the new attribute `recursive` ?
> 
> To the best of my knowledge, a recursive field is a compute field that depends on itself in some capacity in order to perform the computation.
> 
> It's not really new in the sense that the attribute existed prior to this patch, the difference was that it was dynamically set (the ORM did the job of figuring out which fields were recursive and which ones weren't) and now it's done statically (when defining the field). 

:+1: 

> The reason why it's done statically now is because doing it dynamically would rely on components that make the sharing of the field impossible (which reduces bootstrapping performance)

It is still a little sad that it isn't detected automatically since most dev will forget to add this attribute. I may have missed that but is it/would it be possible to add a check on install/update in case someone forget that? I'm pretty sure we have some recursive field on runbot :D