@Elkasitu 2021-04-28
> It is still a little sad that it isn't detected automatically since most dev will forget to add this attribute. I may have missed that but is it/would it be possible to add a check on install/update in case someone forget that? I'm pretty sure we have some recursive field on runbot :D

There is a warning, not sure exactly when it's emitted though